+++ 301/CO_config.h
@@ -70,7 +70,7 @@ extern "C" {
  *
  * This flag is common to multiple configuration macros.
  */
-#define CO_CONFIG_FLAG_CALLBACK_PRE 0x0100
+#define CO_CONFIG_FLAG_CALLBACK_PRE 0x1000
 
 
 /**
@@ -81,7 +81,7 @@ extern "C" {
  *
  * This flag is common to multiple configuration macros.
  */
-#define CO_CONFIG_FLAG_TIMERNEXT 0x0200
+#define CO_CONFIG_FLAG_TIMERNEXT 0x2000
 
 
 /**
@@ -90,22 +90,19 @@ extern "C" {
  * Possible flags, can be ORed:
  * - #CO_CONFIG_FLAG_CALLBACK_PRE - Enable custom callback after preprocessing
  *   received NMT CAN message.
+ *   Callback is configured by CO_NMT_initCallbackPre().
  * - #CO_CONFIG_FLAG_TIMERNEXT - Enable calculation of timerNext_us variable
  *   inside CO_NMT_process().
- *   Callback is configured by CO_NMT_initCallbackPre().
  * - CO_CONFIG_NMT_CALLBACK_CHANGE - Enable custom callback after NMT
  *   state changes. Callback is configured by
  *   CO_NMT_initCallbackChanged().
  * - CO_CONFIG_NMT_MASTER - Enable simple NMT master
- * - CO_CONFIG_NMT_LEDS - Calculate CANopen blinking variables, which can
- *   be used for LEDs.
  */
 #ifdef CO_DOXYGEN
-#define CO_CONFIG_NMT (CO_CONFIG_FLAG_CALLBACK_PRE | CO_CONFIG_FLAG_TIMERNEXT | CO_CONFIG_NMT_CALLBACK_CHANGE | CO_CONFIG_NMT_MASTER | CO_CONFIG_NMT_LEDS)
+#define CO_CONFIG_NMT (CO_CONFIG_FLAG_CALLBACK_PRE | CO_CONFIG_FLAG_TIMERNEXT | CO_CONFIG_NMT_CALLBACK_CHANGE | CO_CONFIG_NMT_MASTER0)
 #endif
 #define CO_CONFIG_NMT_CALLBACK_CHANGE 0x01
 #define CO_CONFIG_NMT_MASTER 0x02
-#define CO_CONFIG_NMT_LEDS 0x04
 
 
 /**
@@ -114,9 +111,9 @@ extern "C" {
  * Possible flags, can be ORed:
  * - #CO_CONFIG_FLAG_CALLBACK_PRE - Enable custom callback after preprocessing
  *   received SDO CAN message.
+ *   Callback is configured by CO_SDO_initCallbackPre().
  * - #CO_CONFIG_FLAG_TIMERNEXT - Enable calculation of timerNext_us variable
  *   inside CO_SDO_process().
- *   Callback is configured by CO_SDO_initCallbackPre().
  * - CO_CONFIG_SDO_SEGMENTED - Enable SDO server segmented transfer.
  * - CO_CONFIG_SDO_BLOCK - Enable SDO server block transfer. If set, then
  *   CO_CONFIG_SDO_SEGMENTED must also be set.
@@ -199,11 +196,17 @@ extern "C" {
  * - #CO_CONFIG_FLAG_TIMERNEXT - Enable calculation of timerNext_us variable
  *   inside CO_TPDO_process().
  * - CO_CONFIG_PDO_SYNC_ENABLE - Enable SYNC object inside PDO objects.
+ * - CO_CONFIG_RPDO_CALLS_EXTENSION - Enable calling configured extension
+ *   callbacks when received RPDO CAN message modifies OD entries.
+ * - CO_CONFIG_TPDO_CALLS_EXTENSION - Enable calling configured extension
+ *   callbacks before TPDO CAN message is sent.
  */
 #ifdef CO_DOXYGEN
-#define CO_CONFIG_PDO (CO_CONFIG_FLAG_CALLBACK_PRE | CO_CONFIG_FLAG_TIMERNEXT | CO_CONFIG_PDO_SYNC_ENABLE)
+#define CO_CONFIG_PDO (CO_CONFIG_FLAG_CALLBACK_PRE | CO_CONFIG_FLAG_TIMERNEXT | CO_CONFIG_PDO_SYNC_ENABLE | CO_CONFIG_RPDO_CALLS_EXTENSION | CO_CONFIG_TPDO_CALLS_EXTENSION)
 #endif
 #define CO_CONFIG_PDO  (CO_CONFIG_RPDO_ENABLE | CO_CONFIG_TPDO_ENABLE)
+#define CO_CONFIG_RPDO_CALLS_EXTENSION 0x02
+#define CO_CONFIG_TPDO_CALLS_EXTENSION 0x04
 
 
 /**
@@ -231,8 +234,8 @@ extern "C" {
  * - #CO_CONFIG_FLAG_TIMERNEXT - Enable calculation of timerNext_us variable
  *   inside CO_SDOclientDownloadInitiate(), CO_SDOclientDownload(),
  *   CO_SDOclientUploadInitiate(), CO_SDOclientUpload().
- * - CO_CONFIG_SDO_CLI_SEGMENTED - Enable SDO server segmented transfer.
- * - CO_CONFIG_SDO_CLI_BLOCK - Enable SDO server block transfer. If set, then
+ * - CO_CONFIG_SDO_CLI_SEGMENTED - Enable SDO client segmented transfer.
+ * - CO_CONFIG_SDO_CLI_BLOCK - Enable SDO client block transfer. If set, then
  *   CO_CONFIG_SDO_CLI_SEGMENTED must also be set.
  * - CO_CONFIG_SDO_CLI_LOCAL - Enable local transfer, if Node-ID of the SDO
  *   server is the same as node-ID of the SDO client. (SDO client is the same
@@ -263,16 +266,50 @@ extern "C" {
 
 
 /**
- * Configuration of LSS master object
+ * Configuration of TIME
  *
  * Possible flags, can be ORed:
  * - #CO_CONFIG_FLAG_CALLBACK_PRE - Enable custom callback after preprocessing
- *   received SDO CAN message.
- *   Callback is configured by CO_LSSmaster_initCallbackPre().
+ *   received TIME CAN message.
+ *   Callback is configured by CO_TIME_initCallbackPre().
  */
 #ifdef CO_DOXYGEN
-#define CO_CONFIG_LSS_MST (CO_CONFIG_FLAG_CALLBACK_PRE)
+#define CO_CONFIG_TIME (CO_CONFIG_FLAG_CALLBACK_PRE)
 #endif
+
+
+/**
+ * Configuration of LEDs object
+ *
+ * Possible flags, can be ORed:
+ * - #CO_CONFIG_FLAG_TIMERNEXT - Enable calculation of timerNext_us variable
+ *   inside CO_NMT_process().
+ * - CO_CONFIG_LEDS_ENABLE - Enable calculation of the CANopen LED indicators.
+ */
+#ifdef CO_DOXYGEN
+#define CO_CONFIG_LEDS (CO_CONFIG_FLAG_TIMERNEXT | CO_CONFIG_LEDS_ENABLE)
+#endif
+#define CO_CONFIG_LEDS_ENABLE 0x01
+
+
+/**
+ * Configuration of LSS objects
+ *
+ * Possible flags, can be ORed:
+ * - #CO_CONFIG_FLAG_CALLBACK_PRE - Enable custom callback after preprocessing
+ *   received CAN message.
+ *   Callback is configured by CO_LSSmaster_initCallbackPre().
+ * - CO_CONFIG_LSS_SLAVE - Enable LSS slave
+ * - CO_CONFIG_LSS_SLAVE_FASTSCAN_DIRECT_RESPOND - Send LSS fastscan respond
+ *   directly from CO_LSSslave_receive() function.
+ * - CO_CONFIG_LSS_MASTER - Enable LSS master
+ */
+#ifdef CO_DOXYGEN
+#define CO_CONFIG_LSS (CO_CONFIG_FLAG_CALLBACK_PRE | CO_CONFIG_LSS_SLAVE | CO_CONFIG_LSS_SLAVE_FASTSCAN_DIRECT_RESPOND | CO_CONFIG_LSS_MASTER )
+#endif
+#define CO_CONFIG_LSS_SLAVE 0x01
+#define CO_CONFIG_LSS_SLAVE_FASTSCAN_DIRECT_RESPOND 0x02
+#define CO_CONFIG_LSS_MASTER 0x10
 
 
 /**
@@ -284,24 +321,61 @@ extern "C" {
  *
  * Possible flags, can be ORed:
  * - CO_CONFIG_GTW_MULTI_NET - Enable multiple network interfaces in gateway
- *   device. This functionality is currntly not implemented.
+ *   device. This functionality is currently not implemented.
  * - CO_CONFIG_GTW_ASCII - Enable gateway device with ASCII mapping (CiA 309-3)
+ * - CO_CONFIG_GTW_ASCII_SDO - Enable SDO client
+ * - CO_CONFIG_GTW_ASCII_NMT - Enable NMT master
+ * - CO_CONFIG_GTW_ASCII_LSS - Enable LSS master
+ * - CO_CONFIG_GTW_ASCII_LOG - Enable non-standard message log read
+ * - CO_CONFIG_GTW_ASCII_ERROR_DESC - Print error description as additional
+ *   comments in gateway-ascii device for SDO and gateway errors.
+ * - CO_CONFIG_GTW_ASCII_PRINT_HELP - use non-standard command "help" to print
+ *   help usage.
+ * - CO_CONFIG_GTW_ASCII_PRINT_LEDS - Display "red" and "green" CANopen status
+ *   LED diodes on terminal.
  */
 #ifdef CO_DOXYGEN
-#define CO_CONFIG_GTW (CO_CONFIG_GTW_MULTI_NET | CO_CONFIG_GTW_ASCII)
+#define CO_CONFIG_GTW (CO_CONFIG_GTW_MULTI_NET | CO_CONFIG_GTW_ASCII | CO_CONFIG_GTW_ASCII_SDO | CO_CONFIG_GTW_ASCII_NMT | CO_CONFIG_GTW_ASCII_LSS | CO_CONFIG_GTW_ASCII_LOG | CO_CONFIG_GTW_ASCII_ERROR_DESC | CO_CONFIG_GTW_ASCII_PRINT_HELP | CO_CONFIG_GTW_ASCII_PRINT_LEDS)
 #endif
 #define CO_CONFIG_GTW_MULTI_NET 0x01
 #define CO_CONFIG_GTW_ASCII 0x02
+#define CO_CONFIG_GTW_ASCII_SDO 0x04
+#define CO_CONFIG_GTW_ASCII_NMT 0x08
+#define CO_CONFIG_GTW_ASCII_LSS 0x10
+#define CO_CONFIG_GTW_ASCII_LOG 0x20
+#define CO_CONFIG_GTW_ASCII_ERROR_DESC 0x40
+#define CO_CONFIG_GTW_ASCII_PRINT_HELP 0x80
+#define CO_CONFIG_GTW_ASCII_PRINT_LEDS 0x100
+
+
+/**
+ * Number of loops of #CO_SDOclientDownload() in case of block download
+ *
+ * If SDO clint has block download in progress and OS has buffer for CAN tx
+ * messages, then #CO_SDOclientDownload() functionion can be called multiple
+ * times within own loop (up to 127). This can speed-up SDO block transfer.
+ */
+#ifdef CO_DOXYGEN
+#define CO_CONFIG_GTW_BLOCK_DL_LOOP 1
+#endif
 
 
 /**
  * Size of command buffer in ASCII gateway object.
  *
- * If large amount of data is transferred (block transfer), the this should be
+ * If large amount of data is transferred (block transfer), then this should be
  * increased to 1000 or more. Buffer may be refilled between the block transfer.
  */
 #ifdef CO_DOXYGEN
-#define CO_CONFIG_GTWA_COMM_BUF_SIZE 100
+#define CO_CONFIG_GTWA_COMM_BUF_SIZE 200
+#endif
+
+
+/**
+ * Size of message log buffer in ASCII gateway object.
+ */
+#ifdef CO_DOXYGEN
+#define CO_CONFIG_GTWA_LOG_BUF_SIZE 2000
 #endif